+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - feat: `NesWindow` uses `NesContainerTheme` instead of Material's card.
  - feat: update mini sprite to improve icons.
  - feat: add `NesIcons.center`.
+ - feat: add ability to customize mouse cursors used by the package.
 
 # 0.12.1
  - fix: theme lerp causing error on null access.